+++ lib/batch.dart
@@ -2,21 +2,26 @@
 // Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// Main feature
 export 'package:batch/src/batch_application.dart';
-export 'package:batch/src/job.dart';
-export 'package:batch/src/step.dart';
-export 'package:batch/src/task.dart';
-export 'package:batch/src/repeat_status.dart' show RepeatStatus;
 
+// Job features
+export 'package:batch/src/job/job.dart';
+export 'package:batch/src/job/step.dart';
+export 'package:batch/src/job/task.dart';
+export 'package:batch/src/job/execution_context.dart';
+export 'package:batch/src/job/job_execution.dart';
+export 'package:batch/src/job/step_execution.dart';
+export 'package:batch/src/job/repeat_status.dart' show RepeatStatus;
+
+// Logging features
 export 'package:batch/src/log_configuration.dart';
 export 'package:batch/src/log/logger_provider.dart';
 export 'package:batch/src/log/log_level.dart';
 export 'package:batch/src/log/input_log_event.dart';
 export 'package:batch/src/log/output_log_event.dart';
-
 export 'package:batch/src/log/filter/log_filter.dart';
 export 'package:batch/src/log/filter/development_log_filter.dart';
 export 'package:batch/src/log/filter/production_log_filter.dart';
-
 export 'package:batch/src/log/output/console_log_output.dart';
 export 'package:batch/src/log/output/file_log_output.dart';