+++ lib/src/batch_application.dart
@@ -3,12 +3,14 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:batch/src/batch_application_impl.dart';
-import 'package:batch/src/job.dart';
+import 'package:batch/src/job/job.dart';
 import 'package:batch/src/log_configuration.dart';
 
 abstract class BatchApplication {
   /// Returns the new instance of [BatchApplication].
-  factory BatchApplication({LogConfiguration? logConfig}) =>
+  factory BatchApplication({
+    LogConfiguration? logConfig,
+  }) =>
       BatchApplicationImpl(logConfig: logConfig);
 
   /// Adds [Job].
@@ -16,6 +18,12 @@ abstract class BatchApplication {
   /// The name of the Job must be unique, and an exception will be raised
   /// if a Job with a duplicate name has already been registered.
   BatchApplication addJob(final Job job);
+
+  /// Adds parameter as global scope.
+  BatchApplication addParameter<T>({
+    required String key,
+    required T value,
+  });
 
   /// Runs this batch application.
   void run();