+++ lib/src/job/job.dart
@@ -3,7 +3,9 @@
 // BSD-style license that can be found in the LICENSE file.
 
 // Project imports:
-import 'package:batch/src/step.dart';
+import 'package:batch/src/parameter/job_relationship.dart';
+import 'package:batch/src/parameter/shared_parameters.dart';
+import 'package:batch/src/job/step.dart';
 
 /// This class represents a job which is the largest unit in batch execution processing.
 ///
@@ -40,6 +42,21 @@ class Job {
     }
 
     steps.add(step);
+
+    return this;
+  }
+
+  Job addParameter<T>({
+    required String key,
+    required T value,
+  }) {
+    SharedParameters.instance.putAsJobScope(
+      key: key,
+      value: value,
+      relationship: JobRelationship(
+        parentJobName: name,
+      ),
+    );
 
     return this;
   }