+++ lib/src/job/step.dart
@@ -3,7 +3,9 @@
 // BSD-style license that can be found in the LICENSE file.
 
 // Project imports:
-import 'package:batch/src/task.dart';
+import 'package:batch/src/parameter/job_relationship.dart';
+import 'package:batch/src/parameter/shared_parameters.dart';
+import 'package:batch/src/job/task.dart';
 
 /// This class represents the processing of each step that constitutes a job in batch processing.
 class Step {
@@ -23,6 +25,22 @@ class Step {
   /// Tasks added by this [nextTask] method are executed in the order in which they are stored.
   Step nextTask(final Task task) {
     tasks.add(task);
+    return this;
+  }
+
+  Step addParameter<T>({
+    required String key,
+    required T value,
+  }) {
+    SharedParameters.instance.putAsStepScope(
+      key: key,
+      value: value,
+      relationship: JobRelationship(
+        parentJobName: '',
+        parentStepName: '',
+      ),
+    );
+
     return this;
   }