+++ lib/src/log/logger_provider.dart
@@ -3,22 +3,22 @@
 // BSD-style license that can be found in the LICENSE file.
 
 // Project imports:
-import 'package:batch/src/log/logger_instance.dart';
+import 'package:batch/src/log/logger.dart';
 
 /// Logging function for trace level.
-final trace = LoggerInstance.instance!.trace;
+final trace = Logger.instance.trace;
 
 /// Logging function for debug level.
-final debug = LoggerInstance.instance!.debug;
+final debug = Logger.instance.debug;
 
 /// Logging function for info level.
-final info = LoggerInstance.instance!.info;
+final info = Logger.instance.info;
 
 /// Logging function for warning level.
-final warning = LoggerInstance.instance!.warning;
+final warning = Logger.instance.warning;
 
 /// Logging function for error level.
-final error = LoggerInstance.instance!.error;
+final error = Logger.instance.error;
 
 /// Logging function for fatal level.
-final fatal = LoggerInstance.instance!.fatal;
+final fatal = Logger.instance.fatal;